Week-2/Day-4/Labs/lab-1/requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.8
Flask-Limiter==3.8.0
PyJWT==2.9.0
